// WYNKWeb/src/app/Models/EmployeeMaster.model.ts
export class Employee_Master {
  EmployeeID: number;
  CMPID: number;
  Title: string;
  FirstName: string;
  MiddleName: string;
  LastName: string;
  Dateofbirth: Date;
  Gender: number;
  PhysicallyChallenged: boolean;
  MaritalStatus: string;
  Category: string;
  DeptCode: number;
  AadhaarNo: string;
  PANNo: string;
  FatherHusbandName: string;
  ReasonForResignation: string;
  DOJ: Date;
  DOR: Date;
  Designation: string;
  EmergencyContactName: string;
  EmergencyContactNo: string;
  BloodGroup: string;
  IsActive: boolean;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: Date;
  UpdatedBy: Date;


}