// WYNKWeb/src/app/Models/EmployeeStatutory.model.ts
export class EmployeeStatutory {

  ID: number;
  CmpID: number;
  EmpID: number;
  IsPFEligible: boolean;
  PFNumber: string;
  PFCommencementDate: Date;
  VPFPercentage: number;
  IsESIEligible: boolean;
  ESINumber: string;
  ESICommencementDate: Date;
  IsProfessionalTaxEligible: boolean;
  GratuityNo: string;
  UANNo: string;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;





}