// WYNKWeb/src/app/Models/EmployeeBank.model.ts
export class EmployeeBank {

  ID: number;
  CmpID: number;
  EmpID: number;
  BankName: string;
  BankBranch: string;
  AccountHoldersName: string;
  AccountType: string;
  AccountNo: string;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
  IFSCCode: string;






}