// WYNKWeb/src/app/Models/EmployeeCommunication.model.ts
export class EmployeeCommunication {

  ID: number;
  CmpID: number;
  EmpID: number;
  PresentAddress1: string;
  PresentAddress2: string;
  PresentLocationID: number;
  PresentLandmark: string;
  PermanentAddress1: string;
  PermanentAddress2: string;
  PermanentLocationID: number;
  PermanentLandmark: string;
  Phone: string;
  MobileNumber1: string;
  MobileNumber2: string;
  EmailID: string;
  AlternateEmailID:string;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;

}